Move Loading to it's own namespace
