+++ Assets/CursedSummit/Loading/GameLoader.cs
@@ -4,11 +4,11 @@ using System.Diagnostics;
 using System.IO;
 using CursedSummit.UI;
 using CursedSummit.Utils;
-using UnityEngine;
 using FindFiles;
+using UnityEngine;
 using Debug = UnityEngine.Debug;
 
-namespace CursedSummit
+namespace CursedSummit.Loading
 {
     /// <summary>
     /// General Game loader (nothing really implemented, only a framework skeleton)
@@ -24,15 +24,16 @@ namespace CursedSummit
         #endregion
 
         #region Constants
-        public const string folderName = "CSData";
+        public const string GameFolderName = "CSData";
         #endregion
 
         #region Fields
         [SerializeField]
         private Progressbar loadingbar;    //Loading bar     
-        private List<ILoader> loaders = new List<ILoader>();
-        private List<FileInfo> allFiles = new List<FileInfo>();
-        private Dictionary<string, List<FileInfo>> extensions = new Dictionary<string, List<FileInfo>>();
+        private readonly List<ILoader> loaders = new List<ILoader>();
+        private readonly List<FileInfo> allFiles = new List<FileInfo>();
+        private readonly Dictionary<string, List<FileInfo>> extensions = new Dictionary<string, List<FileInfo>>();
+        private readonly Dictionary<string, List<FileInfo>> jsonExtensions = new Dictionary<string, List<FileInfo>>();
         #endregion
 
         #region Methods
@@ -62,7 +63,7 @@ namespace CursedSummit
         private IEnumerator Start()
         {
             Stopwatch watch = Stopwatch.StartNew();
-            string localPath = Path.Combine(CSUtils.RootPath, folderName);
+            string localPath = Path.Combine(CSUtils.RootPath, GameFolderName);
             if (!Directory.Exists(localPath))
             {
                 Debug.LogWarning("[GameLoader]: CSData folder could not be located. Creating new one.");
@@ -74,16 +75,29 @@ namespace CursedSummit
                 foreach (FileInfo file in e.Matches())
                 {
                     this.loadingbar.SetLabel("Locating file " + file.FullName);
-
                     this.allFiles.Add(file);
                     List<FileInfo> files;
-                    if (!this.extensions.TryGetValue(file.Extension, out files))
-                    {
-                        files = new List<FileInfo>();
-                        this.extensions.Add(file.Extension, files);
-                    }
 
-                    files.Add(file);
+                    if (file.Extension == ".json")
+                    {
+                        FileInfo jsonFile = new FileInfo(Path.GetFileNameWithoutExtension(file.FullName));
+                        if (!this.jsonExtensions.TryGetValue(jsonFile.Extension, out files))
+                        {
+                            files = new List<FileInfo>();
+                            this.extensions.Add(jsonFile.Extension, files);
+                        }
+                        files.Add(jsonFile);
+                    }
+                    else
+                    {
+                        if (!this.extensions.TryGetValue(file.Extension, out files))
+                        {
+                            files = new List<FileInfo>();
+                            this.extensions.Add(file.Extension, files);
+                        }
+
+                        files.Add(file);
+                    }
                     Debug.Log("[GameLoader]: Located " + file.FullName);
                     yield return null;
                 }
@@ -92,8 +106,7 @@ namespace CursedSummit
             watch.Stop();
             Debug.Log($"[GameLoader]: Found {this.allFiles.Count} files in {watch.Elapsed.TotalSeconds}s");
 
-            watch.Reset();
-            watch.Start();
+            watch.Restart();
             foreach (ILoader loader in this.loaders)
             {
                 Debug.Log("[GameLoader]: Starting loader " + loader.Name);
