+++ Assets/CursedSummit/Loading/JsonLoader.cs.meta
